Fix initialisation of NodesManager
Previously, `NodesManager` registered (passing this) as a event listener before ending initialization in c-tor.
When `onEventDispatch()` is called and this method isn't on UI thread it calls `startUpdatingOnAnimationFrame()`, which in turn uses `mChoreographerCallback`.
`mChoreographerCallback` is set **after** listener is registered which causes NPE in React Choreographer's `postFrameCallback()` method after trying adding callback to the queue.

The simplest solution would be to move registration to the end of the c-tor but Java VM is allowed to reorganize instructions.
Moving construction to static function `createNodesManager()` and subscribing to events after `NodesManager` creation ensures that this won't happen.

Fixes #604.

References:
https://www.ibm.com/developerworks/java/library/j-jtp0618/index.html#2
https://stackoverflow.com/questions/25281301/what-exactly-are-the-dangers-of-passing-this-from-a-java-constructor